# apps/api/src/main/java/com/suncorp/securehub/dto/CommentDto.java
package com.suncorp.securehub.dto;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.LocalDateTime;

@Data @Builder @NoArgsConstructor @AllArgsConstructor
public class CommentDto {
    private Long id;
    private Long requestId;
    private UserDto author;
    private String body;
    private LocalDateTime createdAt;
}