# apps/api/src/main/java/com/suncorp/securehub/dto/CreateRequestDto.java
package com.suncorp.securehub.dto;

import com.suncorp.securehub.entity.SupportRequest.RequestStatus;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import lombok.Data;

@Data
public class CreateRequestDto {
    @NotBlank(message = "Title is required")
    @Size(max = 255, message = "Title must not exceed 255 characters")
    private String title;

    @NotBlank(message = "Description is required")
    private String description;
}