# apps/api/src/main/java/com/suncorp/securehub/dto/LoginRequest.java
package com.suncorp.securehub.dto;

import jakarta.validation.constraints.NotBlank;
import lombok.Data;

@Data
public class LoginRequest {
    @NotBlank(message = "Username is required")
    private String username;

    @NotBlank(message = "Password is required")
    private String password;
}