# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Secure Support Hub API
  description: |
    REST API for the Secure Support Hub — a multi-role support-ticket platform.
    Authentication uses JWT Bearer tokens issued by `POST /api/v1/auth/login`.
  version: 1.0.0
  contact:
    name: Engineering Team

servers:
  - url: http://localhost:8080
    description: Local development

security:
  - BearerAuth: []

components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  schemas:
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          example: user
        password:
          type: string
          format: password
          example: password

    AuthResponse:
      type: object
      properties:
        accessToken:
          type: string
        tokenType:
          type: string
          example: Bearer
        expiresIn:
          type: integer
          example: 86400000
        user:
          $ref: '#/components/schemas/UserDto'

    UserDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        roles:
          type: array
          items:
            type: string
            enum: [USER, TRIAGE, ADMIN]

    SupportRequestDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        title:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        createdByUsername:
          type: string
        assignedToUsername:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    CreateRequestDto:
      type: object
      required: [title, description]
      properties:
        title:
          type: string
          minLength: 3
          maxLength: 200
          example: Billing discrepancy
        description:
          type: string
          minLength: 10
          example: I was charged twice in January.

    UpdateStatusDto:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        assignedToUsername:
          type: string
          nullable: true

    CommentDto:
      type: object
      properties:
        id:
          type: integer
          format: int64
        body:
          type: string
        authorUsername:
          type: string
        createdAt:
          type: string
          format: date-time

    CreateCommentDto:
      type: object
      required: [body]
      properties:
        body:
          type: string
          minLength: 1
          example: I have attached the invoice screenshot.

    ErrorResponse:
      type: object
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
        details:
          type: object
          additionalProperties:
            type: string
        requestId:
          type: string

    PageSupportRequest:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/SupportRequestDto'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

    PageComment:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/CommentDto'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer

paths:
  /api/v1/auth/login:
    post:
      summary: Authenticate and receive JWT
      tags: [Auth]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/v1/me:
    get:
      summary: Get current authenticated user
      tags: [Auth]
      responses:
        '200':
          description: Current user info
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserDto'
        '401':
          description: Not authenticated

  /api/v1/requests:
    get:
      summary: List support requests (paginated)
      tags: [Requests]
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, IN_PROGRESS, RESOLVED, CLOSED]
        - name: assignedTo
          in: query
          schema:
            type: string
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Paginated list of requests
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageSupportRequest'
        '403':
          description: Not authenticated

    post:
      summary: Create a new support request
      tags: [Requests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRequestDto'
      responses:
        '201':
          description: Request created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportRequestDto'
        '400':
          description: Validation error

  /api/v1/requests/{id}:
    get:
      summary: Get a support request by ID
      tags: [Requests]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Support request detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportRequestDto'
        '404':
          description: Not found

    patch:
      summary: Update status / assignment (TRIAGE or ADMIN only)
      tags: [Requests]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateStatusDto'
      responses:
        '200':
          description: Updated request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SupportRequestDto'
        '403':
          description: Insufficient role

  /api/v1/requests/{requestId}/comments:
    get:
      summary: List comments for a request
      tags: [Comments]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: page
          in: query
          schema:
            type: integer
            default: 0
        - name: size
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Paginated comments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PageComment'

    post:
      summary: Add a comment to a request
      tags: [Comments]
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentDto'
      responses:
        '201':
          description: Comment created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentDto'
        '400':
          description: Validation error
        '404':
          description: Request not found